+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.7] - 2021-10-23
 
+### Added
+
+- Added third game
+
 ### Fixed
 
 - Fix title too longer error in GUI.