+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.8] - 2021-10-28
+
+### Added
+
+- Blind effect can now be given in game 3
+
+### Fixed
+
+- Fix when game 3 ends the pvp is still on.
+
 ## [0.1.7] - 2021-10-23
 
 ### Added